requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# utils/image_filter_and_upload.py: 5
# utils/image_utils.py: 3
# utils/run_eppraisal_recently_sold.py: 1
# utils/run_remax_new_listings.py: 4
# utils/run_remax_sold_homes.py: 3
# utils/scraping_utils.py: 3
beautifulsoup4 == 4.6.0

# utils/__init__.py: 1
numpy == 1.14.0

# utils/image_filter_and_upload.py: 11
# utils/image_utils.py: 9
# utils/run_eppraisal_recently_sold.py: 4
# utils/run_remax_new_listings.py: 2,10
# utils/run_remax_sold_homes.py: 2
pandas == 0.22.0

# utils/run_eppraisal_recently_sold.py: 6
# utils/run_remax_new_listings.py: 13,16
# utils/scraping_utils.py: 9
python_dateutil == 2.6.1

# utils/image_filter_and_upload.py: 6,7
# utils/image_utils.py: 4,5
# utils/run_eppraisal_recently_sold.py: 2
# utils/run_remax_new_listings.py: 12
# utils/scraping_utils.py: 5,6
requests == 2.18.4

# utils/run_remax_sold_homes.py: 9
urllib3 == 1.22
